1. Added Single Click Cell Edit option 2. Changed export icons
